requirements.txt
customtkinter
pillow
python-nmap
